feat(tokenizer): implement tokenizer rules for pipes
feat(parser): introduce pipeline rules
feat(execute): implement execution of double pipeline
feat(main): re-enable redirection handling for single commands
feat(pipelines): implement pipe with builtins logic
refactor(pipeline): refactor parser and executor to support multi-command
fix(pipeline): clean up fork failure handling and move redirections into pipeline execution
fix(pipeline): enable redirections on single command
fix(pipeline): free pipeline commands and handle empty or trailing segments
fix(tokenizer): stop word tokens at pipe char
fix(types): use size_t for arg and token indexing
chore(execute): doc pipeline VLA tradeoffs and exit-status TODOs
fix(pipeline): correct redirection handling and cleanup fd ownership

+++ src/argvec.c
@@ -4,7 +4,7 @@
 #include "argvec.h"
 
 void double_argvec_capacity(ArgVec* argv) {
-    int new_cap = argv->cap ? argv->cap * 2 : 8;
+    size_t new_cap = argv->cap ? argv->cap * 2 : 8;
     char** new_args = realloc(argv->args, sizeof(char*) * new_cap);
     if(!new_args) {
         abort(); // Handle memory allocation failure
@@ -35,7 +35,7 @@ void append_arg_end(ArgVec* argv) {
 }
 
 void free_argvec(ArgVec* argv) {
-    for (int i = 0; i < argv->count; i++) {
+    for (size_t i = 0; i < argv->count; i++) {
         free(argv->args[i]);
     }
     free(argv->args);
