+++ src/redirect.h
@@ -1,9 +1,11 @@
 #ifndef REDIRECT_H_
 #define REDIRECT_H_
 
+#include <stdlib.h>
+
 #include "command.h"
 #include "tokenizer.h"
 
-void handle_redirect(Command* command, TokenArray* tokens, int* index, int* start, Token* redirect_token);
+void handle_redirect(Command* command, TokenArray* tokens, size_t* index, size_t* start, Token* redirect_token);
 
 #endif
